+++ pygemmes/_plots/__init__.py
@@ -1,3 +1,1 @@
-
-
 from ._plots import *
