+++ requirements.txt
@@ -4,3 +4,4 @@ scipy
 pyvis
 matplotlib
 opencv-python
+rich[jupyter]
